+++ DocumentBuilder/Runtime/Scripts/VisualElements/DocStyleElement/DSTypeField.cs
@@ -1,11 +1,7 @@
 using NaiveAPI_UI;
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
-using UnityEditor.IMGUI.Controls;
-using UnityEngine;
 using UnityEngine.UIElements;
 namespace NaiveAPI.DocumentBuilder
 {